requirements.txt
fastapi
uvicorn
requests
playwright
pydantic
firecrawl
python-dotenv
PyYAML
google-genai